fix: move core to internal folder
