+++ internal/core/version.go
