+++ tools/config/go-templates/atlas_client.mustache
@@ -11,7 +11,7 @@ import (
 	"strings"
 
 	"github.com/mongodb-forks/digest"
-	"go.mongodb.org/atlas-sdk/core"
+	"go.mongodb.org/atlas-sdk/internal/core"
 )
 
 const (
